EZP-27445: Move editing of drafts to custom controller
